Creating a Config class and its subclasses (#104)
Here, a set of configuration classes is added that allows several dictable file formats to be ingested with common methods and behavior. Some JCSDA borrowed code has been removed as this functionality replaces what was adopted in those modules for this stage of development (treating configuration files as flat templates). 

The name of the namelist management tool was changed to reflect that it is meant to be a templater tool that will handle many types of configuration file templates, not just Fortran namelists. Several variables and methods within the tool were also changed similarly. I the new YAML config class is being used as an alternative to using the user's environment to supply the necessary values.

The test suite was updated to include unit tests for the concrete Config classes. The existing functional tests were modified to test the tool without the use of subprocesses.