Add a method to the config base class to find the depth of a given dictionary.  (#182)
This is part of ticket UW-229 and requires adding a method to the config base class to traverse the dictionary (self.data) and return the depth as an integer. It must increment each time a new dictionary is found as a value and only increment when the structure goes any deeper than the current level.