UW-273: Enable all config file types with templater. (#192)
This should fix an issue found in acceptance testing that doesn't allow an INI file to configure a template. Opens it up to a namelist file also working...technically, but why??

Added tests and removed the old template.py that is deprecated.

Added a newline at the end of the rendered template file.