Apply linting consistently to all code. (#220)
While linting for PR #219 I realized we had a linting bug in our tests. By cd'ing into the src directory to lint uwtools, we weren't automatically using the .pylintrc file for the project. This PR lints the code under uwtools to the standard of the project and updates the test to call pylint on all the code together and from the top level.