Add better error handling. (#219)
Some error handling found during the acceptance testing process was less than fabulous. This is an attempt to do it better for users calling UW Tools via the command line, and also to provide an example of how error handling can be done by Applications that will integrate the tool via Python.

In this example, we catch the specific pyyaml error where it happens, but raise a generic UWConfigError so that we can catch it at the highest level (the caller) and exit the program with a descriptive error. By taking this approach, the integration into a different Python interface will still be able to handle an exception raised by the uwtools package, and handle it however it wants. Our script, set_config.py, chooses to handle it by calling a sys.exit() with the original error.