UW-271 Implement -q/-v args as mutually exclusive (#245)
Made -q and -v mutually exclusive in set_config.py and templater.py. Added logical checks for interactions with the quiet flag and others that don't make sense when included together. Created tests for this change in test_set_config.py and test_templater.py. No changes were made to at parse_to_jinja2.py as -q and -v flags are not available in this file.